CCHMC Stats f
1. The out of the list that you provided we had only 7 matches. I
discussed this with Keith and used the encounter types from our Clarity,
that can possibly be considered as “Face to Face” encounters. I may not
be completely right about that and that’s why I have also included stats
about “excluded” encounter types. You both will be my guides to identify
if the exclusions and inclusions are correct or wrong.
2. The worksheets in the excel are kind of self-explanatory. Any
worksheet with name “Candi” or “Candidate” is the one with possible
“face to face” encounters. Any worksheet with “ Excl” or “Exclusion” has
list and statistics on the encounter types that can be considered not
“Face to Face”. (There could be some that I may have mis-categorized).
3. At CCHMC, we don’t have encounter types such as
“Inpatient”/”Outpatient”/”Emergency”. These are Patient Classes and the
queries to identify these are different. I have added a worksheet
“IP_OP_ED” with this stats.
4. One more thing that I think could be useful to analyze is “PRC”. We
have several “Visit Types” that are categorized as “PRC”s. Something we
can think about if more analysis is needed.