Add PCORnet DDL in various dialects
Add PCORnet CDM generation DDL in various dialects, as well as the
python code required to create it. Various changes to the CDM were made
from its last version, as documented in a new README. Specifically,
TEXT(x) data types are interpreted as VARCHAR(1028) and primary key
constraints are added to all tables.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>