Remove raw_dx and raw_px from pkeys
As discussed in issue 76, remove raw_dx and raw_px from primary keys on
their respective diagnosis and procedure tables. For both tables, this
will mean that data being transformed from PEDSnet/OMOP CDM will need to
be aggregated on the source_value fields. This change was made in order
to more closely adhere to PCORnet specifications and facilitate analysis
code written with those specifications in mind. The various DDL dialect
files are regenerated for this commit.

Fixes #76

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>