Add version links to README
Update the README file to be a little more useful:
* Add links to release branches and ETL Convention documents
* Add link to DDL

IMPORTANT: whenever a new release/tag is created, the links in
README.md must be updated.  Strictly speaking, the vX.Y README.md will
not be part of the vX.Y release, but the network users won't care
about that.